tests/wpt/web-platform-tests/permissions/test-background-fetch-permission.html

<!doctype html>
<meta charset=utf-8>
<title>Test Background Fetch Permission.</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<div id="log"></div>

<script>
  promise_test(function(test) {
    return navigator.permissions.query({name:'background-fetch'}).then(function(result) {
      assert_true(result instanceof PermissionStatus);
    });
  });
</script>
